chore: extract components to other file
